A lot of updates. Mostly data usage optimization. New games, with tests and migrations.

+++ src/contracts/lib/libTicket.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.4.16;
 
 library TicketLib {
     struct Ticket {
-        uint block_number;
-        uint block_time;
+        uint40 block_number;
+        uint32 block_time;
         uint prize;
     }
 }
